requirements.txt
fastapi
uvicorn
pandas
requests
openpyxl
google-genai
python-dotenv
